Fixed level issue

+++ Flexi/Flexi.pyw
@@ -30,14 +30,15 @@ async def sndmsg(ctx, *, msg):
 
 @bot.command()
 async def flex(ctx, level = 0):
-    await ctx.message.delete()
-    channel = bot.get_channel(ctx.channel.id)
-    message = await channel.history(limit=1).flatten()
-    message.reverse()  
+    if level != 0:
+        await ctx.message.delete()
+        channel = bot.get_channel(ctx.channel.id)
+        message = await channel.history(limit=1).flatten()
+        message.reverse()  
 
-    nitroflex = [emoji for emoji in ctx.guild.emojis if "nitroflex" in emoji.name][:level]
+        nitroflex = [emoji for emoji in ctx.guild.emojis if "nitroflex" in emoji.name][:level]
 
-    for emoji in nitroflex:
-        await message[0].add_reaction(emoji)
+        for emoji in nitroflex:
+            await message[0].add_reaction(emoji)
 
 bot.run(TOKEN)
